+++ CHANGELOG.md
@@ -6,8 +6,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 ### Added
+- Batch messages [#185](https://github.com/rokwire/notifications-building-block/issues/185)
 - Add in Airship Push Notifications [#173](https://github.com/rokwire/notifications-building-block/issues/173)
 - Add multiple topic support
+- Add CORS support
 
 ## [1.21.1] - 2024-09-06
 ### Fixed